floating point precision fix; float->double

+++ contrib/brl/bbas/brad/brad_image_atmospherics_est.cxx
@@ -7,13 +7,13 @@
 
 bool brad_estimate_atmospheric_parameters(vil_image_view<float> const& radiance, brad_image_metadata const& mdata, brad_atmospheric_parameters &params)
 {
-   unsigned int ni = radiance.ni();
-   unsigned int nj = radiance.nj();
+  unsigned int ni = radiance.ni();
+  unsigned int nj = radiance.nj();
   // find min and max values in image
   float minval, maxval;
-  vil_math_value_range(radiance,minval,maxval);
+  vil_math_value_range(radiance, minval, maxval);
   // compute histogram for image
-  bsta_histogram<float> h(minval, maxval, 512);
+  bsta_histogram<double> h(minval, maxval, 512);
   for (unsigned j = 0; j<nj; ++j)
     for (unsigned i = 0; i<ni; ++i)
       h.upcount(radiance(i, j), 1.0f);
@@ -41,9 +41,9 @@ bool brad_estimate_atmospheric_parameters(vil_image_view<float> const& radiance,
   unsigned int nj = radiance.nj();
   // find min and max values in image
   float minval, maxval;
-  vil_math_value_range(radiance,minval,maxval);
+  vil_math_value_range(radiance, minval, maxval);
   // compute histogram for image
-  bsta_histogram<float> h(minval, maxval, 512);
+  bsta_histogram<double> h(minval, maxval, 512);
   for (unsigned j = 0; j<nj; ++j)
     for (unsigned i = 0; i<ni; ++i)
       h.upcount(radiance(i, j), 1.0f);
@@ -68,7 +68,7 @@ bool brad_estimate_atmospheric_parameters(vil_image_view<float> const& radiance,
   double skylight = 0.0;
   double deg2rad = vnl_math::pi_over_180;
   double optical_depth = -std::log(vnl_math::pi / (mean_reflectance * mdata.sun_irradiance_ * std::sin(deg2rad*mdata.sun_elevation_)) * (radiance_mean - airlight));
-  optical_depth /= (1.0/std::sin(deg2rad*mdata.view_elevation_) + 1.0/std::sin(deg2rad*mdata.sun_elevation_));
+  optical_depth /= (1.0 / std::sin(deg2rad*mdata.view_elevation_) + 1.0 / std::sin(deg2rad*mdata.sun_elevation_));
 
   if (constrain_atmospheric_params) {
     // Optical depth cannot be less than 0.
@@ -97,7 +97,7 @@ bool brad_estimate_atmospheric_parameters_multi(vil_image_view<float> const& rad
   bands.clear();
 
   for (unsigned p = 0; p < np; p++)
-    bands.push_back(vil_image_view<float>(vil_plane(radiance,p)));
+    bands.push_back(vil_image_view<float>(vil_plane(radiance, p)));
 
   // calculate atmospheric parameters for each band
   for (unsigned p = 0; p < np; p++)
@@ -106,10 +106,10 @@ bool brad_estimate_atmospheric_parameters_multi(vil_image_view<float> const& rad
     float minval, maxval;
     vil_math_value_range(bands[p], minval, maxval);
     // compute histogram for image
-    bsta_histogram<float> h(minval, maxval, 512);
+    bsta_histogram<double> h(minval, maxval, 512);
     for (unsigned j = 0; j < nj; j++)
       for (unsigned i = 0; i < ni; i++)
-        h.upcount(bands[p](i,j), 1.0f);
+        h.upcount(bands[p](i, j), 1.0f);
 
     // compute airlight
     float frac = 0.0001f;
@@ -118,7 +118,7 @@ bool brad_estimate_atmospheric_parameters_multi(vil_image_view<float> const& rad
     double optical_depth = 0.0;
     double skylight = 0.0;
     std::cout << "band " << p << ": min = " << minval << ", airlight = " << airlight
-             << ", optical_depth (fixed) = " << optical_depth << ", skylight (fixed) = " << skylight << std::endl;
+      << ", optical_depth (fixed) = " << optical_depth << ", skylight (fixed) = " << skylight << std::endl;
     // set atmospheric parameters
     params.airlight_multi_.push_back(airlight);
     params.optical_depth_multi_.push_back(optical_depth);
@@ -141,15 +141,15 @@ bool brad_estimate_atmospheric_parameters_multi(vil_image_view<float> const& rad
     float minval, maxval;
     vil_math_value_range(img, minval, maxval);
     // compute histogram for image band p
-    bsta_histogram<float> h(minval, maxval, 512);
+    bsta_histogram<double> h(minval, maxval, 512);
     for (unsigned j = 0; j < nj; j++)
       for (unsigned i = 0; i < ni; i++)
-        h.upcount(img(i,j), 1.0f);
+        h.upcount(img(i, j), 1.0f);
     // compute airlight
     float frac = 0.0001f;
     double airlight = h.value_with_area_below(frac);
     // find image mean
-    float radiance_mean;
+    double radiance_mean;
     vil_math_mean(radiance_mean, img, 0);
     // check the image mean value and the computed airlight
     if (radiance_mean < airlight) {
@@ -160,7 +160,7 @@ bool brad_estimate_atmospheric_parameters_multi(vil_image_view<float> const& rad
     double skylight = 0.0;
     // calculate optical depth
     double deg2rad = vnl_math::pi_over_180;
-    double optical_depth = -std::log(vnl_math::pi / (mean_reflectance*mdata.sun_irradiance_values_[p]*std::sin(deg2rad*mdata.sun_elevation_)) * (radiance_mean-airlight));
+    double optical_depth = -std::log(vnl_math::pi / (mean_reflectance*mdata.sun_irradiance_values_[p] * std::sin(deg2rad*mdata.sun_elevation_)) * (radiance_mean - airlight));
 
     if (constrain_atmospheric_params) {
       // Optical depth cannot be less than 0
@@ -173,7 +173,7 @@ bool brad_estimate_atmospheric_parameters_multi(vil_image_view<float> const& rad
     params.optical_depth_multi_.push_back(optical_depth);
     params.skylight_multi_.push_back(skylight);
     std::cout << "band " << p << ": min = " << minval << ", airlight = " << airlight
-             << ", optical_depth = " << optical_depth << ", skylight (fixed) = " << skylight << std::endl;
+      << ", optical_depth = " << optical_depth << ", skylight (fixed) = " << skylight << std::endl;
   }
   return true;
 
@@ -191,15 +191,15 @@ bool brad_undo_reflectance_estimate(vil_image_view<float> const& reflectance, br
 
   // ideal Lambertian reflector, surface normal = [0 0 1]
   double sun_dot_norm = std::sin(sun_el_rads);
-  double Lsat_horizontal = T_view*(mdata.sun_irradiance_ * sun_dot_norm * T_sun + atm_params.skylight_)/vnl_math::pi + atm_params.airlight_;
+  double Lsat_horizontal = T_view*(mdata.sun_irradiance_ * sun_dot_norm * T_sun + atm_params.skylight_) / vnl_math::pi + atm_params.airlight_;
 
   // convert reflectance values to radiance
   unsigned int ni = reflectance.ni();
   unsigned int nj = reflectance.nj();
-  radiance.set_size(ni,nj);
-  for (unsigned int j=0; j<nj; ++j) {
-    for (unsigned int i=0; i<ni; ++i) {
-      radiance(i,j) = float(reflectance(i,j) * Lsat_horizontal + atm_params.airlight_);
+  radiance.set_size(ni, nj);
+  for (unsigned int j = 0; j<nj; ++j) {
+    for (unsigned int i = 0; i<ni; ++i) {
+      radiance(i, j) = float(reflectance(i, j) * Lsat_horizontal + atm_params.airlight_);
     }
   }
   return true;
@@ -217,18 +217,18 @@ bool brad_estimate_reflectance_image(vil_image_view<float> const& radiance, brad
 
   // ideal Lambertian reflector, surface normal = [0 0 1]
   double sun_dot_norm = std::sin(sun_el_rads);
-  double Lsat_horizontal = T_view*(mdata.sun_irradiance_ * sun_dot_norm * T_sun + atm_params.skylight_)/vnl_math::pi + atm_params.airlight_;
+  double Lsat_horizontal = T_view*(mdata.sun_irradiance_ * sun_dot_norm * T_sun + atm_params.skylight_) / vnl_math::pi + atm_params.airlight_;
 
   // stretch radiance values such that airlight value maps to 0, and ideal reflector maps to 1
   unsigned int ni = radiance.ni();
   unsigned int nj = radiance.nj();
-  reflectance.set_size(ni,nj);
-  for (unsigned int j=0; j<nj; ++j) {
-    for (unsigned int i=0; i<ni; ++i) {
-      double normalized = (radiance(i,j) - atm_params.airlight_) / Lsat_horizontal;
+  reflectance.set_size(ni, nj);
+  for (unsigned int j = 0; j<nj; ++j) {
+    for (unsigned int i = 0; i<ni; ++i) {
+      double normalized = (radiance(i, j) - atm_params.airlight_) / Lsat_horizontal;
       // don't let reflectance value fall below 0
       // value above 1 is unlikely but possible (e.g. specular reflection)
-      reflectance(i,j) = std::max(0.0f, float(normalized));
+      reflectance(i, j) = std::max(0.0f, float(normalized));
     }
   }
   return true;
@@ -248,20 +248,20 @@ bool brad_estimate_reflectance_image_multi(vil_image_view<float> const& radiance
     vil_image_view<float> img = vil_plane(radiance, p);
     vil_image_view<float> reflectance = vil_plane(reflectance_multi, p);
 
-    double T_sun  = std::exp(-atm_params.optical_depth_multi_[p] / std::sin(sun_el_rads));
+    double T_sun = std::exp(-atm_params.optical_depth_multi_[p] / std::sin(sun_el_rads));
     double T_view = std::exp(-atm_params.optical_depth_multi_[p] / std::sin(sat_el_rads));
 
     // ideal Lambertian reflector, surface normal = [0,0,1]
     double sun_dot_norm = std::sin(sun_el_rads);
-    double Lsat_horizontal = T_view*(mdata.sun_irradiance_values_[p]*sun_dot_norm*T_sun + atm_params.skylight_multi_[p])/vnl_math::pi + atm_params.airlight_multi_[p];
+    double Lsat_horizontal = T_view*(mdata.sun_irradiance_values_[p] * sun_dot_norm*T_sun + atm_params.skylight_multi_[p]) / vnl_math::pi + atm_params.airlight_multi_[p];
 
     // stretch radiance values such that airlight value maps to 0, and ideal reflector maps to 1
     for (unsigned j = 0; j < nj; j++) {
       for (unsigned i = 0; i < ni; i++) {
-        double normalized = (img(i,j) - atm_params.airlight_multi_[p]) / Lsat_horizontal;
+        double normalized = (img(i, j) - atm_params.airlight_multi_[p]) / Lsat_horizontal;
         // don't let reflectance value fall below 0
         // value above 1 is unlikely but possible (e.g. specular reflection)
-        reflectance(i,j) = std::min(std::max(0.0f, float(normalized)), 1.0f);
+        reflectance(i, j) = std::min(std::max(0.0f, float(normalized)), 1.0f);
       }
     }
   }
